+++ Globals.cpp
@@ -27,5 +27,5 @@ const std::string Globals::kDummy = "DummySymbol";
 const std::vector<std::string> Globals::kPersistenceDelimiters = {":::", "{--->", "|||", "}--->"};
 
 const float Globals::kBaseProb = 0.0001;
-const unsigned int Globals::kBandGap = 5;
-const unsigned short Globals::kMaxMultiWordSize = 4;
+const uint8_t Globals::kBandGap = 5;
+const uint8_t Globals::kMaxMultiWordSize = 4;
